+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Web Scraper** Reduced webscraper delay from 5 seconds to 3 seconds
 - **Package Handler** Added further replacement values for Arch Gitlab packages (example gjs)
+- **Config Handler** Added sphinx docstrings
 
 ### Bug fixes